requirements.txt
customtkinter==5.2.2
pillow==10.0.1
python-nmap==0.7.1
